+++ examples/ai-core/src/embed-many/mistral.ts
@@ -1,8 +1,6 @@
 import { mistral } from '@ai-sdk/mistral';
 import { embedMany } from 'ai';
-import dotenv from 'dotenv';
-
-dotenv.config();
+import 'dotenv/config';
 
 async function main() {
   const { embeddings, usage } = await embedMany({