+++ examples/ai-core/src/stream-object/groq.ts
@@ -1,9 +1,7 @@
 import { createOpenAI } from '@ai-sdk/openai';
 import { streamObject } from 'ai';
-import dotenv from 'dotenv';
+import 'dotenv/config';
 import { z } from 'zod';
-
-dotenv.config();
 
 const groq = createOpenAI({
   apiKey: process.env.GROQ_API_KEY ?? '',