+++ examples/ai-core/src/stream-object/openai-array.ts
@@ -1,26 +1,23 @@
 import { openai } from '@ai-sdk/openai';
 import { streamObject } from 'ai';
-import dotenv from 'dotenv';
+import 'dotenv/config';
 import { z } from 'zod';
 
-dotenv.config();
-
 async function main() {
-  const result = await streamObject({
-    model: openai('gpt-4o-2024-08-06', { structuredOutputs: true }),
+  const { elementStream: destinations } = await streamObject({
+    model: openai('gpt-4o'),
     output: 'array',
     schema: z.object({
-      name: z.string(),
-      class: z
-        .string()
-        .describe('Character class, e.g. warrior, mage, or thief.'),
+      city: z.string(),
+      country: z.string(),
       description: z.string(),
+      attractions: z.array(z.string()).describe('List of major attractions.'),
     }),
-    prompt: 'Generate 3 hero descriptions for a fantasy role playing game.',
+    prompt: 'What are the top 5 cities for short vacations in Europe?',
   });
 
-  for await (const hero of result.elementStream) {
-    console.log(hero);
+  for await (const destination of destinations) {
+    console.log(destination); // destination is a complete array element
   }
 }