+++ examples/ai-core/src/stream-text/amazon-bedrock.ts
@@ -1,8 +1,6 @@
 import { bedrock } from '@ai-sdk/amazon-bedrock';
 import { streamText } from 'ai';
-import dotenv from 'dotenv';
-
-dotenv.config();
+import 'dotenv/config';
 
 async function main() {
   const result = await streamText({