+++ examples/ai-core/src/stream-text/google-vertex-fullstream.ts
@@ -1,10 +1,8 @@
 import { vertex } from '@ai-sdk/google-vertex';
 import { streamText } from 'ai';
-import dotenv from 'dotenv';
+import 'dotenv/config';
 import { z } from 'zod';
 import { weatherTool } from '../tools/weather-tool';
-
-dotenv.config();
 
 async function main() {
   const result = await streamText({