+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/openai
 
+## 0.0.61
+
+### Patch Changes
+
+- 8132a60: feat (provider/openai): support reasoning token usage and max_completion_tokens
+
 ## 0.0.60
 
 ### Patch Changes